Setup for performance tests reporting bot (#63)
* add webhook

* moving performance tests to a different case

* only run performance tests during performance test

* fix filter typo

* clean up experimentation

* move filter off of build

* kick

* name steps

* yaml

* yaml

* update linuxmain

* attempt 2

* cleanup

* cleanup

* remove performance test from workflow

* point to bot-gh.vapor.codes

* added swift-version to gitignore
