+++ .gitignore
@@ -5,4 +5,4 @@ Packages
 Package.pins
 Package.resolved
 DerivedData
-
+.swift-version
